pyproject.toml
[project]
name = "dmi-mlverif-notebooks"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "jupyter-book",
    "papermill",
    "pandas",
    "requests",
    "python-dotenv",
    "gitpython",
    "loguru",
    "pyyaml",
    "isodate>=0.7.2",
    "geopandas>=1.0.1",
    "xarray>=2024.7.0",
    "bidict>=0.23.1",
    "owslib>=0.31.0",
    "inflection>=0.5.1",
    "zarr>=2.18.2",
    "dask>=2024.8.0",
    "s3fs>=2025.3.2",
    "scipy>=1.13.1",
    "matplotlib>=3.9.4",
]

[project.scripts]
trigger_mlverif_build = "dmi_mlverif_notebooks.trigger_workflow:main"
create_dummy_data = "dmi_mlverif_notebooks.create_dummy_data:main"
book_setup = "dmi_mlverif_notebooks.book_setup:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
